test/requirements.txt
psycopg2-binary==2.9.9
pytest==7.4.0
pytest-postgresql==5.0.0
python-dotenv==1.0.0 